android-tfa: move entity to right folder
