+++ android/DFM/src/main/kotlin/com.darakeon.dfm/api/DfmService.kt
@@ -1,7 +1,7 @@
 package com.darakeon.dfm.api
 
 import com.darakeon.dfm.api.entities.Body
-import com.darakeon.dfm.api.entities.TFA
+import com.darakeon.dfm.api.entities.tfa.TFA
 import com.darakeon.dfm.api.entities.accounts.AccountList
 import com.darakeon.dfm.api.entities.extract.Extract
 import com.darakeon.dfm.api.entities.login.Login
